%x <= unsigned int

+++ src/lang/type/t.c
@@ -9,5 +9,5 @@ main(int argc, const char *argv[]) {
   float f0 = 3.14f;
   double f1 = 2.7182;
   printf("pi %f stored as %8x\n", f0, *(unsigned int *)&f0);
-  printf("e %f stored as %8x\n", f1, *(unsigned long*)&f1);
+  printf("e %f stored as %8x\n", f1, *(unsigned int*)&f1);
 }
